requirements.txt
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
numpy
opencv-python-headless
Pillow
psycopg2
requests==2.28.1
pika
pymongo==4.4.0
marshmallow